Create symlinks for dotfiles if they don't already exist.

+++ install.rb
@@ -5,14 +5,12 @@
 home = File.expand_path('~')
 
 Dir['*'].each do |file|
-  next if file =~ /install/ || file =~ /README/
+  next if file =~ /install/ ||
+          file =~ /README/
   if file =~ /^[A-Z]/
     target = File.join(home, file)
   else
     target = File.join(home, ".#{file}")
   end
-  `ln -ns #{File.expand_path file} #{target}`
+  `ln -ns #{File.expand_path file} #{target}` unless File.exist?(target) || File.symlink?(target)
 end
-
-# `git submodule sync`
-# `git submodule update --init --recursive`
